% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSII_Code.R
\name{rar.rm}
\alias{rar.rm}
\title{Rare species suppression}
\usage{
rar.rm(tableau.AD, n)
}
\arguments{
\item{tableau.AD}{a community data matrix}

\item{n}{a minima number of occurence}
}
\value{
the same community data matrix but without the species with an occurence below n
}
\description{
Suppress species that don't have a sufficient number of occurence
}
