% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSII_Code.R
\name{structure.plotV2}
\alias{structure.plotV2}
\title{structure plot version 2}
\usage{
structure.plotV2(INDICE, FACTOR, MULTI = T, MTITLE = "", ABMAX = 5,
  col1 = "grey60", col2 = "white", col3 = "red", col4 = "orange",
  noms = "T", cex_noms = 1, erreur = "sem", w_err = 1, sp_star = 1,
  adj_meth = "BH", stars = "T", BASE = T, ...)
}
\arguments{
\item{INDICE}{An object issued from \code{\link{ComStructIndices}} function}

\item{FACTOR}{A factor list, a barplot of species mean abundances will be performed for each factor level. If no factor is specified, MULTI=F should be specified.}

\item{MULTI}{If no factor is specified, MULTI=F should be specified}

\item{MTITLE}{Main title of the plot}

\item{ABMAX}{Numerical value of the maximum abundance}

\item{col1}{Colour information for the Reference mean abundances barplot}

\item{col2}{Colour information for the Reference mean abundances in assessed community barplot, i.e. "missing abundances"}

\item{col3}{Colour information for the abundances of target species in the assessed community}

\item{col4}{Colour information for the "higher abundances" in the assessed community}

\item{noms}{If other than "T", species names are not given}

\item{cex_noms}{expansion factor for species names}

\item{erreur}{(facultative) error calculation \code{"sem"} (default) is standard error of the mean \code{"IC"} will calculate confidence interval at 5 percent \code{"var"} or \code{"sd"} or any other R function can be use (but beware, you can't use the \code{na.rm=T} argument..).}

\item{w_err}{error bar width}

\item{sp_star}{width between stars and error bars}

\item{adj_meth}{p adjust method: \code{"bond"},\code{"BH"},\code{"hoch"}, etc. or \code{"none"} (then p correspond to a wilcoxon test between reference and modality)}

\item{stars}{"T" (default), draw or note the stars, anything but "T" does not draw the stars}

\item{BASE}{T (default), draw or not the reference community for each modality}

\item{...}{any arguments from a \code{\link{barplot}} function}
}
\description{
same as \code{\link{structure.plot}} but with some more arguments (especially error bars)
}
\details{
Be careful, the p adjustment is done only within a modality, perhaps that would be better to adjust it through all the modalitties (to be modified later). The warnings come from ties, but it seems it does not change the outcome of tests...
}
\seealso{
\code{\link{ComStructIndices}} et \code{\link{structure.plot}}
}
